[#48] 파일 분리
